some progress on integrating kafka
