+++ ActivityAnalysisService/activityanalysisservice/src/main/resources/main.properties
@@ -1,0 +1,2 @@
+kafka.bootstrap.servers = localhost:9192
+kafka.topic = test_topic
